Move DeltaEnergyOne, Zero, and related traits to hoomd-interaction.
