+++ hoomd-interaction/src/single.rs
@@ -71,3 +71,54 @@ where
         self.0.site_energy(site_properties)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use hoomd_microstate::boundary::Open;
+    use hoomd_microstate::property::{Point, Position};
+    use hoomd_microstate::{Body, Microstate};
+    use hoomd_vector::Cartesian;
+    use rstest::*;
+
+
+        struct TestSE;
+
+        impl<S> SiteEnergy<S> for TestSE
+        where
+            S: Position<Vector = Cartesian<2>>,
+        {
+            fn site_energy(&self, site_properties: &S) -> f64 {
+                site_properties.position()[0] + site_properties.position()[1]
+            }
+        }
+
+        #[fixture]
+        fn microstate() -> Microstate<Point<Cartesian<2>>, Point<Cartesian<2>>, Open> {
+            let mut microstate = Microstate::new();
+            microstate
+                .extend_bodies([
+                    Body::point(Cartesian::from([1.0, 0.0])),
+                    Body::point(Cartesian::from([-1.0, 3.0])),
+                ])
+                .expect("hard-coded bodies should be in the boundary");
+            microstate
+        }
+
+        #[rstest]
+        fn single_total(microstate: Microstate<Point<Cartesian<2>>, Point<Cartesian<2>>, Open>) {
+            let test_se = TestSE;
+            let single = Single(test_se);
+
+            assert_eq!(single.total_energy(&microstate), 3.0);
+        }
+
+        #[rstest]
+        fn single_site(microstate: Microstate<Point<Cartesian<2>>, Point<Cartesian<2>>, Open>) {
+            let test_se = TestSE;
+            let single = Single(test_se);
+
+            assert_eq!(single.site_energy(&microstate.sites()[0].properties), 1.0);
+            assert_eq!(single.site_energy(&microstate.sites()[1].properties), 2.0);
+        }
+}
